+++ chess/king.py
@@ -5,3 +5,6 @@ class King(Piece):
     white_str ="♚"
     black_str = "♔" 
 
+#agregar movientos ortogonales y diagonales
+#distancia mov ort
+#agregar distancia para las diagonales