+++ chess/pawn.py
@@ -2,5 +2,7 @@ from chess.pieces import Piece      #PEON CLASE
 
 class Pawn(Piece):
 
-    white_str = "♜"
-    black_str = "♖"
+    white_str = "♙" 
+    black_str = "♟"
+
+