+++ chess/queen.py
@@ -3,4 +3,9 @@ from chess.pieces import Piece         #REINA CLASE
 class Queen(Piece):
     
     white_str = "♛"
-    black_str = "♕"  
+    black_str = "♕"
+    
+
+
+#tengo movimientos ortogonales
+#agregar movimientos diagonales