+++ tests/testPeon.py
@@ -1,0 +1,13 @@
+
+
+
+#movimiento peon negro inicial, movimiento peon blanco inicial
+#movimientos no iniciales (codigo primera ver)
+#comer piezas (diagonales) para ambos colores
+#funcion comer y mover
+# extend para unir metodos
+# no se puede mover si hay una pieza adelante
+
+
+
+#promote = hace que se convierta el peon al llegar al final, se cambie por una pieza